refactor: moves Auditable class to common/domain
