+++ src/main/java/com/lp/v2/domains/auth/AuthController.java
@@ -74,6 +74,4 @@ public class AuthController {
     }
 
 
-
-
 }
